added draw functionality and input assertion for new grid creation

+++ script.js
@@ -12,27 +12,67 @@ function createCards(gridSize) {
     }
 }
 
+// Create new grid with input size
+function newGrid() {
+    let newGridSize = parseInt(prompt("Please enter how many squares per side (1-100):"),10);
+    if(isNaN(newGridSize) | newGridSize > 100 | newGridSize < 1) {
+        alert("Please enter a valid number between 1 and 100")
+        return;
+    }
+
+    // Delete old grid
+    deleteGrid();    
+    // Create new grid with new size
+    createCards(newGridSize);
+}
+
+// Delete cards from DOM
 function deleteGrid() {
     while(cardContainer.hasChildNodes())
         cardContainer.removeChild(cardContainer.lastChild);
 }
 
 
+//--------------------------------------------------------------
+// Definitions
 DEFAULT_GRID_SIZE = 16;
 
 const cardContainer = document.querySelector('.card-container');
 const btnContainer = document.querySelector('.btn-container');
 const gridButton = document.querySelector('#grid-btn');
+const toggleButton = document.querySelector('#toggle-btn');
+const clearButton = document.querySelector('#clear-btn');
 
-// initialize grid
+
+// Initialize grid
 createCards(DEFAULT_GRID_SIZE);
 
-gridButton.addEventListener('click', () => {
-    let newGridSize = prompt("Please enter how many squares per side (this will clear the current grid!):")
+// Create new grid with input size
+gridButton.addEventListener('click', newGrid);
 
-    // Delete old grid
-    deleteGrid();    
+// Draw on mouseover
+cardContainer.addEventListener('mouseover', e => {
+    let card = document.querySelector('.card');
+    card = e.target;
+    if(card.classList.contains('card-container')) return;
+    card.setAttribute('style', 'background-color: black');
+});
 
-    // Create new grid with new size
-    createCards(newGridSize);
-})
+
+// Clear canvas button
+clearButton.addEventListener('click', () => {
+    const cards = document.querySelectorAll('.card');
+    cards.forEach(card => {
+        card.setAttribute('style', 'background-color: white');
+    });
+});
+
+// Toggle grid button 
+// toggleButton.addEventListener('click', () => {
+//     let cards = query.querySelectorAll('.card');
+
+//     cards.forEach(card => {
+        
+//     })
+
+// })
